Added sparkline to landing page table
