+++ src/components/Sparkline/sparkline.styled.js
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div``;
