+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Navbar } from "./Navbar";
+export { default as Sparkline } from "./Sparkline";
