+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as Coin } from "./Coin";
+export { default as CoinList } from "./CoinList";
+export { default as Portfolio } from "./Portfolio";
